requirements.txt
numpy~=2.0.0
pandas~=2.2.2
scikit-learn~=1.5.0
pytz~=2024.1
joblib~=1.4.2
PyYAML~=6.0.1
openpyxl~=3.1.5